refactor: receive bitcoin and ethereum components name and location
