+++ src/frontend/src/icp/components/transactions/IcTransactions.svelte
@@ -11,7 +11,7 @@
 	import type { IcToken, IcTransactionUi } from '$icp/types/ic';
 	import { token } from '$lib/derived/token.derived';
 	import { loadNextTransactions } from '$icp/services/ic-transactions.services';
-	import IcReceiveBitcoin from '$icp/components/receive/IcReceiveBitcoin.svelte';
+	import IcTransactionsBitcoinStatus from '$icp/components/transactions/IcTransactionsBitcoinStatusBalance.svelte';
 	import Info from '$icp/components/info/Info.svelte';
 	import { WALLET_PAGINATION } from '$icp/constants/ic.constants';
 	import type { ComponentType } from 'svelte';
@@ -26,7 +26,7 @@
 	import { slide } from 'svelte/transition';
 	import IcTransactionsCkEthereumListeners from '$icp/components/transactions/IcTransactionsCkEthereumListeners.svelte';
 	import { nullishSignOut } from '$lib/services/auth.services';
-	import IcReceiveEthereum from '$icp/components/receive/IcReceiveEthereum.svelte';
+	import IcTransactionsEthereumStatus from '$icp/components/transactions/IcTransactionsEthereumStatus.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
 	import Header from '$lib/components/ui/Header.svelte';
 	import IcTokenModal from '$icp/components/tokens/IcTokenModal.svelte';
@@ -85,9 +85,9 @@
 
 	<svelte:fragment slot="end">
 		{#if $tokenCkBtcLedger}
-			<IcReceiveBitcoin />
+			<IcTransactionsBitcoinStatus />
 		{:else if ckEthereum}
-			<IcReceiveEthereum />
+			<IcTransactionsEthereumStatus />
 		{/if}
 	</svelte:fragment>
 </Header>
