+++ src/frontend/src/icp/components/transactions/IcTransactionsBitcoinStatusProgress.svelte
